+++ CHANGE_LOG.md
@@ -5,14 +5,19 @@ releases on the way from an old version to a new one.  Fix any deprecation warni
 release before upgrading to the next one.  The documentation next to each Deprecated annotation
 tells you what to use instead.  Once we delete the deprecated methods, that documentation goes too.
 
-## Release 3.7.2 2021-12-20: Fix RrbTree.split()
+## Release 3.8.0 2021-12-28: BaseList.appendSome(() -> Option<E>)
+- Replaced BaseList.appendWhen(Fn0<Boolean> test, E item) with appendSome(Fn0<Option<E>>).
+  This is a better design because it can do everything the first design does, but
+  allows you to avoid constructing a new E unless you need it.
+
+### Release 3.7.2 2021-12-20: Fix RrbTree.split()
 - Fixed Bug #47: "RrbTree.split() returns wrong type" and added test to prevent regressions.
   Thank you, @fcurts, for reporting this!
 
-### Release 3.7.1 2021-12-20
+#### Release 3.7.1 2021-12-20
 - Added BaseList.appendWhen() for fluent list building with optional additions.
 
-### Release 3.7.0 2021-12-19
+#### Release 3.7.0 2021-12-19
  - RRB Tree .join() method appears to work after removing strict-node optimization.
    Many thanks to hours of conversations with Andy Fingerhut @jafingerhut and for his
    great unit tests and fixes.  Without his help, this would not have happened.